+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Version 2.2.0
+* Configurable fields in Chronicle backend #101
+* Improve batch-id (UUID) performance #108
+
 ## Version 2.1.0
 * Make configuration updateable from tests - #98
 * Make wrapped authorizer backend configurable - #104